abstract out ExternalData type
* removes support for createMergedResolver and ExtendSchema transform in favor of visitResult, simplifying the refactoring.
