+++ packages/delegate/src/defaultMergedResolver.ts
@@ -1,9 +1,11 @@
 import { defaultFieldResolver, GraphQLResolveInfo } from 'graphql';
 
-import { getResponseKeyFromInfo, getErrors } from '@graphql-tools/utils';
+import { getResponseKeyFromInfo } from '@graphql-tools/utils';
 
 import { handleResult } from './results/handleResult';
 import { getSubschema } from './Subschema';
+import { getErrors, isExternalData } from './externalData';
+import { ExternalData } from './types';
 
 /**
  * Resolver that knows how to:
@@ -12,7 +14,7 @@ import { getSubschema } from './Subschema';
  * c) handle external to internal enum coversion
  */
 export function defaultMergedResolver(
-  parent: Record<string, any>,
+  parent: ExternalData,
   args: Record<string, any>,
   context: Record<string, any>,
   info: GraphQLResolveInfo
@@ -22,16 +24,16 @@ export function defaultMergedResolver(
   }
 
   const responseKey = getResponseKeyFromInfo(info);
-  const errors = getErrors(parent, responseKey);
 
   // check to see if parent is not a proxied result, i.e. if parent resolver was manually overwritten
   // See https://github.com/apollographql/graphql-tools/issues/967
-  if (!errors) {
+  if (!isExternalData(parent)) {
     return defaultFieldResolver(parent, args, context, info);
   }
 
   const result = parent[responseKey];
   const subschema = getSubschema(parent, responseKey);
+  const errors = getErrors(parent, responseKey);
 
   return handleResult(result, errors, subschema, context, info);
 }
