+++ packages/delegate/src/index.ts
@@ -1,11 +1,11 @@
 export { delegateToSchema, delegateRequest } from './delegateToSchema';
 export { createRequestFromInfo, createRequest } from './createRequest';
 export { defaultMergedResolver } from './defaultMergedResolver';
-export { createMergedResolver } from './createMergedResolver';
 export { handleResult } from './results/handleResult';
 export { Subschema, isSubschema, isSubschemaConfig, cloneSubschemaConfig, getSubschema } from './Subschema';
 
 export * from './applySchemaTransforms';
 export * from './delegationBindings';
+export * from './externalData';
 export * from './transforms';
 export * from './types';
