+++ packages/delegate/src/symbols.ts
@@ -1,2 +1,3 @@
+export const ERROR_SYMBOL = Symbol('subschemaErrors');
 export const OBJECT_SUBSCHEMA_SYMBOL = Symbol('initialSubschema');
 export const FIELD_SUBSCHEMA_MAP_SYMBOL = Symbol('subschemaMap');
